Migrate PhotoPermissions to Factory and use for toolbar item
